+++ src/main/angular/lab-app/src/app/model/Hit.ts
